executor: split test and create hash_join_test to improve ci
