+++ executor/test/jointest/hashjoin/BUILD.bazel
@@ -1,23 +1,23 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_test(
-    name = "jointest_test",
-    timeout = "moderate",
+    name = "hashjoin_test",
+    timeout = "short",
     srcs = [
-        "join_test.go",
+        "hash_join_test.go",
         "main_test.go",
     ],
     flaky = True,
     race = "on",
-    shard_count = 41,
+    shard_count = 13,
     deps = [
         "//config",
         "//meta/autoid",
         "//planner/core",
         "//session",
+        "//sessionctx/variable",
         "//testkit",
-        "//util",
-        "//util/memory",
+        "//util/dbterror/exeerrors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//tikv",
